feat: support building + releasing core24 snaps (#38)
* feat: Add support for `platforms` key

core24 uses the `platforms` key for specifying the architecture of the snap
rather than `architectures`. This commit adds a conditional that properly
handles the `platforms` key when determining which arches to build for.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>

* feat: Add support for `--platform` flag

`core24` supports using the platform flag with the Launchpad
remote-builders. This commit adds a conditional that checks if the
base is core24. If the base is core24, it uses the `--platfrom` flag
rather than the build-for flag.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>

---------

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>