feat(release-to-candidate): Support projects that doesn't specify the `version` property
It is possible that the Snapcraft project doesn't specify the `version`
property in the snapcraft.yaml, instead setting the snap version string
via the `craftctl set version=_version` mechanism.  In this case the
parse-snapcraft-yaml job will set the version variable to `null`, making
the release-to-candidate job unable to detect the built snap due to
filename mismatch.

This patch enhances the release-to-candidate job so that the built
snap will still be matched via filename matching.

Signed-off-by: 林博仁(Buo-ren Lin) <buo.ren.lin@gmail.com>